Click and drag cards
Cards in the player's hand can be clicked and dragged now, but it's still buggy and is being refined. Currently, dragging any but the last rendered card will cause the last rendered card to disappear and will show a duplicate of the card being dragged in the hand. Also, the selected card flickers. Only the last rendered card functions as intended.